VERSION 0.4.0 now works again
In version 0.3.0 I tried to detect SIGTRAP, too.
This made it no more work in all situations.

So I decided to refactor the inner logic completely
according to the ptrace documentation, which, btw.
sucks a lot and also seems to be wrong here and there.

Because now I get spurious SIGTRAP when running slowdown.
This was not noted by the previous versions, because
they were not able to distinguish between ptrace() based
SIGTRAP and other ones.  But now I observe a non-ptrace-TRAP
being delivered to the forked child once.

I really have no idea why.